feat(runbotics-orchestrator-ui): :sparkles: Changes gateway validation
